refactor ProductListItemComponent and ProductGridItemComponent
